requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
httpx==0.28.1
redis==5.2.1
pydantic==2.10.6
python-multipart==0.0.20
numpy==1.26.4
torch==2.6.0
transformers==4.48.3
huggingface-hub==0.36.2
sentence-transformers==3.4.1
chromadb==0.6.3
langchain==0.3.19
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
watchdog==6.0.0
PyPDF2==3.0.1
python-docx==1.1.2
langid==1.1.6
deep-translator==1.11.4
googletrans-py==4.0.0
gTTS==2.5.4
google-genai==1.66.0
SQLAlchemy==2.0.46
psycopg[binary]==3.2.9
alembic==1.14.1
